Back a few stages
This is a perl script rather than a module but it roughly works as a starting point. Running the test prints out the first column and creates a new csv called `new.csv`.

The parameter of the CSV location that I'm passing in in the test is being ignored, plus the formatting is shonky, but it's reading in the CSV and operating on it.