Add folder for output
Seems likely that there will be two phases of parsing the data; the first stage is categorising and checking to make sure the data is as expected, the second is pulling various things out of it. Create this folder to store intermediate output, and ignore it as we do not want to commit the actual data, just the framework.