First steps towards pulling out the time
I am hampered in my desire to search for better solutions by lack of internet,
as detailed in the comments (for now...). This is a first stab at a very
re-inventing the wheel way of pulling out the times. For now it just fakes the
time and puts it into th new spreadsheet.

Lack of internet also prevents me testing this as I don't have all the
dependencies locally and my Vagrant VM is not running. Tests will certainly fail
at this point and I have not begun updating them yet.