Rewrite a test using Test::Spec
Coming back to this project after a long hiatus showed that unit tests are not enough documentation to let me know where I am; I need to know what they are testing.

A more useful module for testing might be [Test:Spec](https://metacpan.org/pod/Test::Spec) which allows you to define the context of a test and have this output printed, so running the tests, even when successful, gives you an idea of what you are testing.

This commit adds Test::Spec and also rewrites one test using it.

Two things to note:

  - `use Test::Spec` automatically turns on strict and warnings
  - if you end with `done_testing()` the output is not printed; you need to `runtests` (`runtests unless caller` is a useful idiom described in the docs)