Move all tests into the context
This commit doesn't change much but will make it easier to see changes in subequent commits.

"Reading the input" is also not a great name for the context but that illustrates the problem I'm attempting to address; it's hard to remember what I was doing/planning. When I'm back in the mindset of this project hopefully I will come up with a better name.